+++ opentelemetry/src/metrics/noop.rs
@@ -9,6 +9,8 @@ use crate::{
 };
 use std::sync::Arc;
 
+#[cfg(feature = "experimental_metrics_bound_instruments")]
+use super::instruments::BoundSyncInstrument;
 use super::instruments::SyncInstrument;
 
 /// A no-op instance of a `MetricProvider`
@@ -59,8 +61,25 @@ impl NoopSyncInstrument {
     }
 }
 
-impl<T> SyncInstrument<T> for NoopSyncInstrument {
+impl<T: Send + Sync + 'static> SyncInstrument<T> for NoopSyncInstrument {
     fn measure(&self, _value: T, _attributes: &[KeyValue]) {
+        // Ignored
+    }
+
+    #[cfg(feature = "experimental_metrics_bound_instruments")]
+    fn bind(&self, _attributes: &[KeyValue]) -> Box<dyn BoundSyncInstrument<T> + Send + Sync> {
+        Box::new(NoopBoundSyncInstrument { _private: () })
+    }
+}
+
+#[cfg(feature = "experimental_metrics_bound_instruments")]
+struct NoopBoundSyncInstrument {
+    _private: (),
+}
+
+#[cfg(feature = "experimental_metrics_bound_instruments")]
+impl<T> BoundSyncInstrument<T> for NoopBoundSyncInstrument {
+    fn measure(&self, _measurement: T) {
         // Ignored
     }
 }